+++ src/main/kotlin/com/zepben/auth/common/AuthException.kt
@@ -1,3 +1,3 @@
-package com.zepben.auth
+package com.zepben.auth.common
 
 class AuthException(val statusCode: Int, message: String? = null): Exception(message)