+++ changelog.md
@@ -1,10 +1,13 @@
 ### v1.3.0
+# TODO: UPDATE THIS
 
 ##### Breaking Changes
-* None.
+* Split server and client specific code into two packages:
+  * `com.zepben.auth.server`
+  * `com.zepben.auth.client`
 
 ##### New Features
-* None.
+* Added client-side authentication class, `ZepbenAuthenticator`
 
 ##### Enhancements
 * None.